Add two new functions: avlbst_clone() and avlbst_defrag()
avlbst_clone() is for cloning the avlbst from a node. Can be used for bucket sorting algorithm.

avlbst_defrag() perform the defragment to the avlbst to make the memory layout more compact by move the nodes to lower addresses, and collect the previously freed memory to a larger block which can contain larger data.

Signed-off-by: 0xaa55 <0xaa55@0xaa55.com>